Suggest #RStats hashtag
Feel free to tweak wording on this if you like. To avoid overlap with other... less savory hashtags (some people share their grooming practices on Tuesdays, apparently), it might be helpful to encourage people to tag with both #TidyTuesday and #RStats to make it easier to find relevant results (particularly for classes). The other stuff isn't overwhelming, but some of it is NSFW, so [#TidyTuesday #rstats](https://twitter.com/search?q=(%23TidyTuesday%20AND%20%23rstats)&f=live) is probably safer.