Copy/paste-able checklists
This still isn't ideal. It's difficult to easily see the difference between "I did not check this" and "I checked this and the answer was no." I'd rather give the people the ability to cleanly tell us "this is good", though, so I *think* this is about as good as it can be for the moment. More iterations to come!