Instructions for reviewing datasets. (#691)
* Instructions for reviewing datasets.

I intentionally don't link to this from anywhere yet. I intend to send people directly to this page via social media/Slack. If it's helpful, perhaps we'll promote it more. I can also imagine a relatively simple Shiny app that could streamline this process (applying labels directly, probably requiring a DSLC Slack login to access it to keep out spam + promote the community).

* Copy/paste-able checklists

This still isn't ideal. It's difficult to easily see the difference between "I did not check this" and "I checked this and the answer was no." I'd rather give the people the ability to cleanly tell us "this is good", though, so I *think* this is about as good as it can be for the moment. More iterations to come!