Lowercase 'pydytuesday' in readme.md
I could not figure out why the python snippet would not work for me, even after 'pip install pydytuesday' was succesful.

It was a matter of lower casing the module name. 

I'm not a senior Python dev, and I don't know if this is specific to my setup, but if it's not, then maybe you can lowercase the readme script so other people don't get tripped up.