Use tempdir to avoid path issues.
When I re-ran cleaning, some entries changed. I did this because I saw weird formatting in one entry (which ended up to just be the comma-separated lists of otitle), but it looks like the data was updated since you pulled it. It all looks fine but let me know if you see any strangeness after I accept this!