Chess player ratings (#921)
* Add cleaning.R

* Add intro.md

* Add meta.yaml

* Add fide_ratings_august.csv

* Add fide_ratings_september.csv

* Add fide_ratings_august.md

* Add fide_ratings_september.md

* Add September-2025-Open.png

* Apply suggestions from code review

* Use tempdir to avoid path issues.

When I re-ran cleaning, some entries changed. I did this because I saw weird formatting in one entry (which ended up to just be the comma-separated lists of otitle), but it looks like the data was updated since you pulled it. It all looks fine but let me know if you see any strangeness after I accept this!

* Accept submission

* Update data/2025/2025-09-23/meta.yaml

---------

Co-authored-by: Jon Harmon <jonthegeek@gmail.com>
Co-authored-by: jonthegeek <jonthegeek@users.noreply.github.com>